test(memoService): add unit tests for memoService functionality
- Implemented tests to verify memo creation, validation in memoService.

+++ src/main/java/com/Layla/Oh/Simple/Memo/Manager/repository/JpaMemoRepository.java
@@ -21,7 +21,7 @@ public class JpaMemoRepository implements MemoRepository {
     }
 
     @Override
-    public Optional<Memo> findById(int id) {
+    public Optional<Memo> findById(Long id) {
         Memo memo = em.find(Memo.class, id);
         return Optional.ofNullable(memo);
     }
